requirements.txt
alembic==1.0.3
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
passlib==1.7.1
pycparser==2.19
python-dateutil==2.7.5
python-editor==1.0.3
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2018.7
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.14
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
